Added Support for on demand initialization of WorkManager instead of on startup
