+++ core/src/main/java/com/rudderstack/android/sdk/core/RudderFlushConfig.java
@@ -5,13 +5,16 @@ import com.google.gson.annotations.SerializedName;
 import java.io.Serializable;
 
 public class RudderFlushConfig implements Serializable {
-    static final String RUDDER_FLUSH_CONFIG_FILE_NAME = "RudderFlushConfig";
     @SerializedName("dataPlaneUrl")
     String dataPlaneUrl;
     @SerializedName("authHeaderString")
     String authHeaderString;
     @SerializedName("anonymousHeaderString")
     String anonymousHeaderString;
+    @SerializedName("logLevel")
+    int logLevel;
+    @SerializedName("flushQueueSize")
+    int flushQueueSize;
 
     public String getDataPlaneUrl() {
         return dataPlaneUrl;
@@ -29,13 +32,15 @@ public class RudderFlushConfig implements Serializable {
         return flushQueueSize;
     }
 
-    @SerializedName("flushQueueSize")
-    int flushQueueSize;
+    public int getLogLevel() {
+        return logLevel;
+    }
 
-    public RudderFlushConfig(String dataPlaneUrl, String authHeaderString, String anonymousHeaderString, int flushQueueSize) {
+    public RudderFlushConfig(String dataPlaneUrl, String authHeaderString, String anonymousHeaderString, int flushQueueSize, int logLevel) {
         this.dataPlaneUrl = dataPlaneUrl;
         this.authHeaderString = authHeaderString;
         this.anonymousHeaderString = anonymousHeaderString;
         this.flushQueueSize = flushQueueSize;
+        this.logLevel = logLevel;
     }
 }
